+++ mempool/errors.go
@@ -8,4 +8,5 @@ var (
 	ErrExpectedOneError   = errors.New("expected 1 error")
 	ErrNotEVMTransaction  = errors.New("transaction is not an EVM transaction")
 	ErrNonceGap           = errors.New("tx nonce is higher than account nonce")
+	ErrNonceLow           = errors.New("tx nonce is lower than account nonce")
 )